refactor(web): do not use StorageClient in InstallationFinished
